Rejiggering some things.
Make it possible for users to change their email (need to investigate verification for this; current idea is to send two emails out, with a unique verification link in each). This means each user gets a UUID.

Add a private UUID to pending user records. This will be the new user's record when verified.

Probably want this UUID to survive through transient -> pending -> registered. This way permissions assigned to a user carry over through transitions.

UI tweaks to month view.